requirements.txt
# Core framework
fastapi==0.115.12
fastapi-security==0.1.0
uvicorn==0.21.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-jose==3.5.0
pymongo==4.13.2
python-multipart==0.0.20
numpy==1.26.4

# Agent framework
openai==1.75.0
langchain==0.3.23
langchain-openai==0.2.9
langchain-community==0.3.21
langgraph==0.2.72
langsmith==0.3.45

# Connectors
duckduckgo_search==8.1.1

google-api-core==2.25.0
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# Email service
resend==2.12.0

# Testing
pytest==8.3.5
pytest-asyncio==0.22.0
pytest-mock==3.12.0
httpx